Add fake titlebar to theme tile
